Ensure that changes to input .proto files are watched
Without this, the build system won't notice that a .proto file has been
changed and will not invoke the builder again.

Tested this in the example/ directory with `dart run build_runner watch`
and making changes to the `proto/my_proto.proto` file and watching the
resulting `.dart` files be regenerated.